+++ tests/TestRunner/app/Marshalling/RecordTests.js
@@ -269,258 +269,6 @@ describe(module.id, function () {
         expect(struct.dbl).toBe(1.67);
     });
 
-    it("simd_float2", function(){
-        var f = getFloat2();
-        expect(f[0].toFixed(4)).toBe(1.2345.toFixed(4));
-        expect(f[1].toFixed(4)).toBe(2.3456.toFixed(4));
-        var fi = incrementFloat2(f);
-        expect(fi[0].toFixed(4)).toBe(2.2345.toFixed(4));
-        expect(fi[1].toFixed(4)).toBe(3.3456.toFixed(4));
-
-    });
-
-    it("simd_float3", function(){
-        var f = getFloat3();
-        expect(f[0].toFixed(4)).toBe(1.2345.toFixed(4));
-        expect(f[1].toFixed(4)).toBe(2.3456.toFixed(4));
-        expect(f[2].toFixed(4)).toBe(3.4567.toFixed(4));
-        var fi = incrementFloat3(f);
-        expect(fi[0].toFixed(4)).toBe(2.2345.toFixed(4));
-        expect(fi[1].toFixed(4)).toBe(3.3456.toFixed(4));
-        expect(fi[2].toFixed(4)).toBe(4.4567.toFixed(4));
-    });
-
-    it("simd_float4", function(){
-        var f = getFloat4();
-        expect(f[0].toFixed(4)).toBe(1.2345.toFixed(4));
-        expect(f[1].toFixed(4)).toBe(2.3456.toFixed(4));
-        expect(f[2].toFixed(4)).toBe(3.4567.toFixed(4));
-        expect(f[3].toFixed(4)).toBe(4.5678.toFixed(4));
-        var fi = incrementFloat4(f);
-        expect(fi[0].toFixed(4)).toBe(2.2345.toFixed(4));
-        expect(fi[1].toFixed(4)).toBe(3.3456.toFixed(4));
-        expect(fi[2].toFixed(4)).toBe(4.4567.toFixed(4));
-        expect(fi[3].toFixed(4)).toBe(5.5678.toFixed(4));
-    });
-
-    it("simd_double2", function(){
-        var d = getDouble2();
-        expect(d[0].toFixed(4)).toBe(1.2345.toFixed(4));
-        expect(d[1].toFixed(4)).toBe(2.3456.toFixed(4));
-        var di = incrementDouble2(d);
-        expect(di[0].toFixed(4)).toBe(2.2345.toFixed(4));
-        expect(di[1].toFixed(4)).toBe(3.3456.toFixed(4));
-    });
-
-    it("simd_double3", function(){
-        var d = getDouble3();
-        expect(d[0].toFixed(4)).toBe(1.2345.toFixed(4));
-        expect(d[1].toFixed(4)).toBe(2.3456.toFixed(4));
-        expect(d[2].toFixed(4)).toBe(3.4567.toFixed(4));
-        var di = incrementDouble3(d);
-        expect(di[0].toFixed(4)).toBe(2.2345.toFixed(4));
-        expect(di[1].toFixed(4)).toBe(3.3456.toFixed(4));
-        expect(di[2].toFixed(4)).toBe(4.4567.toFixed(4));
-    });
-
-    it("simd_double4", function(){
-        var d = getDouble4();
-        expect(d[0].toFixed(4)).toBe(1.2345.toFixed(4));
-        expect(d[1].toFixed(4)).toBe(2.3456.toFixed(4));
-        expect(d[2].toFixed(4)).toBe(3.4567.toFixed(4));
-        expect(d[3].toFixed(4)).toBe(4.5678.toFixed(4));
-        var di = incrementDouble4(d);
-        expect(di[0].toFixed(4)).toBe(2.2345.toFixed(4));
-        expect(di[1].toFixed(4)).toBe(3.3456.toFixed(4));
-        expect(di[2].toFixed(4)).toBe(4.4567.toFixed(4));
-        expect(di[3].toFixed(4)).toBe(5.5678.toFixed(4));
-
-    });
-
-    it("SCNVector3ToFloat3", function() {
-        var v = _SCNVector3ToFloat3({x: 1.23, y: 2.3456, z:3.4567});
-        expect(v[0].toFixed(4)).toBe((1.23).toFixed(4));
-        expect(v[1].toFixed(4)).toBe((2.3456).toFixed(4));
-        expect(v[2].toFixed(4)).toBe((3.4567).toFixed(4));
-    });
-
-    it("SCNVector4ToFloat4", function() {
-        var v = _SCNVector4ToFloat4({x: 1.23, y: 2.3456, z:3.4567, w: 4.5678});
-        expect(v[0].toFixed(4)).toBe((1.23).toFixed(4));
-        expect(v[1].toFixed(4)).toBe((2.3456).toFixed(4));
-        expect(v[2].toFixed(4)).toBe((3.4567).toFixed(4));
-        expect(v[3].toFixed(4)).toBe((4.5678).toFixed(4));
-    });
-
-    it("SCNVector3FromFloat3", function() {
-        var v = _SCNVector3FromFloat3(getFloat3());
-        expect(v.x.toFixed(4)).toBe((1.2345).toFixed(4));
-        expect(v.y.toFixed(4)).toBe((2.3456).toFixed(4));
-        expect(v.z.toFixed(4)).toBe((3.4567).toFixed(4));
-    });
-
-    it("SCNVector4FromFloat4", function() {
-        var v = _SCNVector4FromFloat4(getFloat4());
-        expect(v.x.toFixed(4)).toBe((1.2345).toFixed(4));
-        expect(v.y.toFixed(4)).toBe((2.3456).toFixed(4));
-        expect(v.z.toFixed(4)).toBe((3.4567).toFixed(4));
-        expect(v.w.toFixed(4)).toBe((4.5678).toFixed(4));
-    });
-
-    it("simd_float4x4Matrix", function(){
-       var simdMatrix = getMatrixFloat4x4();
-       for (var i = 0; i < 16; i++) {
-         expect(simdMatrix.columns[i%4][Math.floor(i/4)].toFixed(4)).toBe((i*3.1415).toFixed(4));
-       }
-    });
-
-    it("simd_float4x3Matrix", function(){
-       var simdMatrix = getMatrixFloat4x3();
-       for (var i = 0; i < 12; i++) {
-         expect(simdMatrix.columns[i%4][Math.floor(i/4)].toFixed(4)).toBe((i*3.1415).toFixed(4));
-       }
-    });
-
-    it("simd_float4x2Matrix", function(){
-       var simdMatrix = getMatrixFloat4x2();
-       for (var i = 0; i < 8; i++) {
-         expect(simdMatrix.columns[i%4][Math.floor(i/4)].toFixed(4)).toBe((i*3.1415).toFixed(4));
-       }
-    });
-
-    it("simd_float3x4Matrix", function(){
-       var simdMatrix = getMatrixFloat3x4();
-       for (var i = 0; i < 12; i++) {
-         expect(simdMatrix.columns[i%3][Math.floor(i/3)].toFixed(4)).toBe((i*3.1415).toFixed(4));
-       }
-    });
-
-    it("simd_float3x3Matrix", function(){
-       var simdMatrix = getMatrixFloat3x3();
-       for (var i = 0; i < 9; i++) {
-         expect(simdMatrix.columns[i%3][Math.floor(i/3)].toFixed(4)).toBe((i*3.1415).toFixed(4));
-       }
-    });
-
-    it("simd_float3x2Matrix", function(){
-       var simdMatrix = getMatrixFloat3x2();
-       for (var i = 0; i < 6; i++) {
-         expect(simdMatrix.columns[i%3][Math.floor(i/3)].toFixed(4)).toBe((i*3.1415).toFixed(4));
-       }
-    });
-
-    it("simd_float2x4Matrix", function(){
-       var simdMatrix = getMatrixFloat2x4();
-       for (var i = 0; i < 8; i++) {
-         expect(simdMatrix.columns[i%2][Math.floor(i/2)].toFixed(4)).toBe((i*3.1415).toFixed(4));
-       }
-    });
-
-    it("simd_float2x3Matrix", function(){
-       var simdMatrix = getMatrixFloat2x3();
-       for (var i = 0; i < 6; i++) {
-         expect(simdMatrix.columns[i%2][Math.floor(i/2)].toFixed(4)).toBe((i*3.1415).toFixed(4));
-       }
-    });
-
-    it("simd_float2x2Matrix", function(){
-       var simdMatrix = getMatrixFloat2x2();
-       for (var i = 0; i < 4; i++) {
-         expect(simdMatrix.columns[i%2][Math.floor(i/2)].toFixed(4)).toBe((i*3.1415).toFixed(4));
-       }
-    });
-
-    it("simd_double4x4Matrix", function(){
-        var simdMatrix = getMatrixDouble4x4();
-        for (var i = 0; i < 16; i++) {
-          expect(simdMatrix.columns[i%4][Math.floor(i/4)].toFixed(4)).toBe((i*3.1415).toFixed(4));
-        }
-     });
-
-     it("simd_double4x3Matrix", function(){
-        var simdMatrix = getMatrixDouble4x3();
-        for (var i = 0; i < 12; i++) {
-          expect(simdMatrix.columns[i%4][Math.floor(i/4)].toFixed(4)).toBe((i*3.1415).toFixed(4));
-        }
-     });
-
-     it("simd_double4x2Matrix", function(){
-        var simdMatrix = getMatrixDouble4x2();
-        for (var i = 0; i < 8; i++) {
-          expect(simdMatrix.columns[i%4][Math.floor(i/4)].toFixed(4)).toBe((i*3.1415).toFixed(4));
-        }
-     });
-
-     it("simd_double3x4Matrix", function(){
-        var simdMatrix = getMatrixDouble3x4();
-        for (var i = 0; i < 12; i++) {
-          expect(simdMatrix.columns[i%3][Math.floor(i/3)].toFixed(4)).toBe((i*3.1415).toFixed(4));
-        }
-     });
-
-     it("simd_double3x3Matrix", function(){
-        var simdMatrix = getMatrixDouble3x3();
-        for (var i = 0; i < 9; i++) {
-          expect(simdMatrix.columns[i%3][Math.floor(i/3)].toFixed(4)).toBe((i*3.1415).toFixed(4));
-        }
-     });
-
-     it("simd_double3x2Matrix", function(){
-        var simdMatrix = getMatrixDouble3x2();
-        for (var i = 0; i < 6; i++) {
-          expect(simdMatrix.columns[i%3][Math.floor(i/3)].toFixed(4)).toBe((i*3.1415).toFixed(4));
-        }
-     });
-
-     it("simd_double2x4Matrix", function(){
-        var simdMatrix = getMatrixDouble2x4();
-        for (var i = 0; i < 8; i++) {
-          expect(simdMatrix.columns[i%2][Math.floor(i/2)].toFixed(4)).toBe((i*3.1415).toFixed(4));
-        }
-     });
-
-     it("simd_double2x3Matrix", function(){
-        var simdMatrix = getMatrixDouble2x3();
-        for (var i = 0; i < 6; i++) {
-          expect(simdMatrix.columns[i%2][Math.floor(i/2)].toFixed(4)).toBe((i*3.1415).toFixed(4));
-        }
-     });
-
-     it("simd_double2x2Matrix", function(){
-        var simdMatrix = getMatrixDouble2x2();
-        for (var i = 0; i < 4; i++) {
-          expect(simdMatrix.columns[i%2][Math.floor(i/2)].toFixed(4)).toBe((i*3.1415).toFixed(4));
-        }
-     });
-
-     it("SCNMatrix4FromMat4", function() {
-        const m1 = getMatrixFloat4x4();
-        const m2 = _SCNMatrix4FromMat4(m1);
-
-        for (let col = 0; col < 4; col++) {
-            for (let row = 0; row < 4; row++) {
-                expect(m2[`m${col+1}${row+1}`].toFixed(4)).toBe((m1.columns[col][row]).toFixed(4));
-            }
-        }
-    });
-
-    it("SCNMatrix4ToMat4", function() {
-        const m1 = {};
-        for (let col = 0; col < 4; col++) {
-            for (let row = 0; row < 4; row++) {
-                m1[`m${col+1}${row+1}`] = 3.1415*(row*4 + col);
-            }
-        }
-
-        const m2 = _SCNMatrix4ToMat4(m1);
-
-        for (let col = 0; col < 4; col++) {
-            for (let row = 0; row < 4; row++) {
-                expect((m2.columns[col][row]).toFixed(4)).toBe(m1[`m${col+1}${row+1}`].toFixed(4));
-            }
-        }
-    });
-
     // TODO
     // it("ComplexStruct", function() {
     //     var object = {
